+++ trunk/source/WorldServer/Packets/OP_AllCharactersDescReplyMsg_Packet.h
@@ -7,7 +7,7 @@
 
 class OP_AllCharactersDescReplyMsg_Packet : public EQ2Packet {
 public:
-	OP_AllCharactersDescReplyMsg_Packet(uint16_t version)
+	OP_AllCharactersDescReplyMsg_Packet(uint32_t version)
 		: EQ2Packet(version) {
 		RegisterElements();
