+++ trunk/source/WorldServer/Packets/OP_AllWSDescRequestMsg_Packet.h
@@ -8,7 +8,7 @@
 
 class OP_AllWSDescRequestMsg_Packet : public EQ2Packet {
 public:
-	OP_AllWSDescRequestMsg_Packet(uint16_t version)
+	OP_AllWSDescRequestMsg_Packet(uint32_t version)
 		: EQ2Packet(version) {
 		
 	}
