+++ trunk/source/WorldServer/Packets/OP_DeleteCharacterRequestMsg_Packet.h
@@ -13,7 +13,7 @@ extern WorldDatabase database;
 
 class OP_DeleteCharacterRequestMsg_Packet : public EQ2Packet {
 public:
-	OP_DeleteCharacterRequestMsg_Packet(uint16_t version)
+	OP_DeleteCharacterRequestMsg_Packet(uint32_t version)
 		: EQ2Packet(version) {
 		RegisterElements();
