+++ trunk/source/WorldServer/Packets/OP_LsClientBaselogReplyMsg_Packet.h
@@ -4,7 +4,7 @@
 
 class OP_LsClientBaselogReplyMsg_Packet : public ClientLogReply_Packet {
 public:
-	OP_LsClientBaselogReplyMsg_Packet(uint16_t version)
+	OP_LsClientBaselogReplyMsg_Packet(uint32_t version)
 		: ClientLogReply_Packet(version) {
 		name = "Base Log";
 	}
