+++ trunk/source/WorldServer/Packets/OP_LsClientCrashlogReplyMsg_Packet.h
@@ -4,7 +4,7 @@
 
 class OP_LsClientCrashlogReplyMsg_Packet : public ClientLogReply_Packet {
 public:
-	OP_LsClientCrashlogReplyMsg_Packet(uint16_t version)
+	OP_LsClientCrashlogReplyMsg_Packet(uint32_t version)
 		: ClientLogReply_Packet(version) {
 		name = "Crash Log";
 	}
