+++ trunk/source/WorldServer/Packets/OP_LsClientVerifylogReplyMsg_Packet.h
@@ -4,7 +4,7 @@
 
 class OP_LsClientVerifylogReplyMsg_Packet : public ClientLogReply_Packet {
 public:
-	OP_LsClientVerifylogReplyMsg_Packet(uint16_t version)
+	OP_LsClientVerifylogReplyMsg_Packet(uint32_t version)
 		: ClientLogReply_Packet(version) {
 		name = "Verify Log";
 	}
