+++ trunk/source/WorldServer/Packets/OP_WorldListMsg_Packet.h
@@ -6,7 +6,7 @@
 
 class OP_WorldListMsg_Packet : public EQ2Packet {
 public:
-	OP_WorldListMsg_Packet(uint16_t version)
+	OP_WorldListMsg_Packet(uint32_t version)
 		: EQ2Packet(version) {
 		RegisterElements();
