+++ trunk/source/common/Packets/EQ2Packet.h
@@ -11,8 +11,8 @@ class EQ2Packet : public Packet {
 	friend OpcodeManager;
 
 public:
-	void SetVersion(uint16_t new_version) { Version = new_version; }
-	uint16_t GetVersion() { return Version; }
+	void SetVersion(uint32_t new_version) { Version = new_version; }
+	uint32_t GetVersion() { return Version; }
 	uint16_t GetOpcode() { return opcode; }
 	uint8_t PreparePacket(uint16_t MaxLen);
 	uint32_t serialize(unsigned char *dest) const;
@@ -26,11 +26,11 @@ public:
 	bool EQ2Compressed;
 
 protected:
-	EQ2Packet(uint16_t version);
+	EQ2Packet(uint32_t version);
 
 	uint8_t app_opcode_size;
 	uint16_t opcode;
 
 private:
-	uint16_t Version;
+	uint32_t Version;
 };
